Made a search function that can find names
Made a search bar with some search functionality. Currently it initially shows names, then narrows it down when you use the search bar. As of now, it has uses an example json with names and their emails. This can be converted to player names along with their most well-known team. Still it'll be easy to convert these.

I am figuring out if there is a manual way I can insert data. But as of now, it can grab json data.