Fixed intermittent infinite loading bug (I think)
I'm not honestly sure why I did what I did works. I explained it within the modified file, but I am assuming a service worker for parsing the table data needs to be created on each re-generation of the table data and/or chart data so I removed the condition for this. It probably is less memory efficient, but this appears (at least) to fix the problem we were having.